Refactor data types to use 64-bit integers for memory management and process handling

+++ common/common.h
@@ -36,11 +36,11 @@
 typedef int bool;
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
-typedef unsigned int uint32_t;
+typedef unsigned int uint64_t;
 typedef unsigned long long uint64_t;
-typedef uint32_t size_t;
-typedef uint32_t paddr_t;
-typedef uint32_t vaddr_t;
+typedef uint64_t size_t;
+typedef uint64_t paddr_t;
+typedef uint64_t vaddr_t;
 
 /* putchar is provided by each side: kernel uses SBI Console Putchar,
  * user provides its own (syscall-based) implementation. */
